加入班级层次，课程仓库和作业区改造基本完毕

+++ src/main/java/com/kwang/study/homework/controller/HomeworkController.java
@@ -1,14 +1,16 @@
 package com.kwang.study.homework.controller;
 
 import com.kwang.study.auth.utils.AuthenticationUserUtil;
+import com.kwang.study.auth.utils.UserInfoUtils;
 import com.kwang.study.common.R;
 import com.kwang.study.homework.dto.request.HomeworkCreateDTO;
-import com.kwang.study.homework.pojo.Homework;
 import com.kwang.study.homework.pojo.HomeworkDetail;
 import com.kwang.study.homework.service.HomeworkService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.util.Assert;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -16,6 +18,7 @@ import org.springframework.web.multipart.MultipartFile;
 import javax.validation.Valid;
 import java.io.IOException;
 import java.util.List;
+import java.util.Objects;
 
 import static com.kwang.study.constant.ApiPrefixConstant.HOMEWORK_BASE_PREFIX;
 
@@ -26,6 +29,7 @@ import static com.kwang.study.constant.ApiPrefixConstant.HOMEWORK_BASE_PREFIX;
 public class HomeworkController {
 
     private final HomeworkService homeworkService;
+    private final UserInfoUtils userInfoUtils;
 
     // --- Teacher Endpoints ---
 
@@ -36,9 +40,12 @@ public class HomeworkController {
      * 文件字段：files
      */
     @PostMapping("/publish")
-    @PreAuthorize("hasRole('ROLE_TEACHER')")
     public ResponseEntity<R<HomeworkDetail>> publishHomework(@Valid @RequestPart("dto") HomeworkCreateDTO dto,
                                                              @RequestPart(value = "files", required = false) List<MultipartFile> smallFiles) throws IOException {
+        // NOTE: 目前只有教师才可以操作，管理员是否有权限待商榷
+        if (!userInfoUtils.currentUserInClassIsTeacher())
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         // 获取认证身份
         dto.setTeacherId(AuthenticationUserUtil.getCurrentUserId());
         HomeworkDetail createdHomework = homeworkService.createHomework(dto, smallFiles);
@@ -49,8 +56,10 @@ public class HomeworkController {
      * 教师查看自己发布的作业列表
      */
     @GetMapping("/teacher/all")
-    @PreAuthorize("hasRole('ROLE_TEACHER')")
     public ResponseEntity<R<List<HomeworkDetail>>> getTeacherHomeworks() {
+        if (!userInfoUtils.currentUserInClassIsTeacher())
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         Long teacherId = AuthenticationUserUtil.getCurrentUserId();
         List<HomeworkDetail> homeworks = homeworkService.getHomeworksByTeacher(teacherId);
         return ResponseEntity.ok(R.success(homeworks));
@@ -71,8 +80,17 @@ public class HomeworkController {
      * @return 操作结果
      */
     @DeleteMapping("/{homeworkId}")
-    @PreAuthorize("hasRole('ROLE_TEACHER')")
     public ResponseEntity<R<Void>> deleteHomework(@PathVariable Long homeworkId) {
+        // NOTE: 教师和管理员可以删除作业
+        if (!(userInfoUtils.currentUserInClassIsTeacher() ||
+                AuthenticationUserUtil.currentUserIsAdmin()))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
+        HomeworkDetail homework = homeworkService.getHomeworkById(homeworkId);
+        // 只可以删除自己发布的作业
+        if (!Objects.equals(homework.getTeacherId(), AuthenticationUserUtil.getCurrentUserId()))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         homeworkService.deleteHomework(homeworkId);
         return ResponseEntity.ok(R.success(null));
     }
@@ -80,11 +98,21 @@ public class HomeworkController {
     // --- Student Endpoints ---
 
     /**
-     * 学生查看所有作业列表
+     * 学生查看所有作业列表（即学生所在班级的所有作业列表）
      */
-    @GetMapping("/all")
+    @GetMapping("/student/all")
     public ResponseEntity<R<List<HomeworkDetail>>> getAllHomeworks() {
-        List<HomeworkDetail> homeworks = homeworkService.getAllHomeworksForStudent();
+        List<HomeworkDetail> homeworks = homeworkService.getAllHomeworksForStudent(AuthenticationUserUtil.getCurrentUserName());
         return ResponseEntity.ok(R.success(homeworks));
     }
+
+    /**
+     * 查看某个班级的所有作业
+     */
+    @GetMapping("/class/{classId}")
+    public ResponseEntity<R<List<HomeworkDetail>>> getHomeworksByClassId(@PathVariable Long classId) {
+        List<HomeworkDetail> result = homeworkService.getAllHomeworksInClass(classId);
+        return ResponseEntity.ok(R.success(result));
+    }
+
 }
